Fix clipping targets
**Problem**

Clipping system is designed to apply the clipping only the entities
that are not masking system targets as optimization.
But currently due to a bug the entities that are masking system
targets can be also the targets of the clipping system.

This problem can cause a bad performance impact because clippings
are done more than needed.

**Change for solution**

Fix the bug.

In the change for the bug fix, introducing
Object3D.userData.isEntityObject3DRoot to distinguish whether an
Three.js Object3D is a root Object3D of MR entity to handle clipping
targets properly.