Fix Masking System
**Problem**

Perhaps due to the recent changes to mapping and clipping systems,
background of mr div entities seem not to be masked.

The root issue seems to be in masking system. Child objects
associated with an entity need to be masked but we have been missing
to set up their materials for stencil.

**Change for fix**

Traverse the objects associated with an entity and set up their
materials.

This commit introduces a new traverseObjects method to MREntity
that runs the passed through function on the objects associated with
an entity as a helper method because such traversal is needed
in masking and clipping systems, and it would be used from
more many places in the future.