Introduce BoundaryVisibilitySystem
**Problem**

Currently entities outside of their parent panels are masked with
stencil or clipped with clipping. But stencil and clipping are
costly.

**Change for solution**

Introduce a new system named BoundaryVisibilitySystem that makes
the entities that outside of their parent panels invisible with
`.object3D.visible = false`. It should be much cheaper than stencil
or clipping.