Fix MREntity.traverse()
**Problem**

MREntity.traverse() wants to go down to a child only if the child
is MREntity. But due to a wrong check, it always goes down.

**Change for solution**

Fix the bug. Currently `((!child) instanceof MREntity)` is used
to decide not to go down but it would be always false because
`!child` would be boolean (true/false). `!(child instanceof MREntity)`
would be what it really needs to be.

Also, this commit includes a minor clean up.