Introduce simple mr-stats
**Problem**

"stats=true" attribute in mr-app can show the stats panel on the
page but it won't be shown in immersive mode. We need stats panel
that can be visible even in immersive mode for performance test.

We need fast stats panel because stats panel that can affect
performance doesn't really make sense.

**Change for solution**

Introduce MRStatsEntity (mr-stats). For simplicity and performance
it is based on MRTextEntity and just shows the FPS counter number
as text.